[BE] feat: update, delete community

+++ community-service/src/main/java/com/msa/community_service/domain/community/service/CommunityServiceImpl.java
@@ -3,6 +3,7 @@ package com.msa.community_service.domain.community.service;
 
 
 import com.msa.community_service.domain.community.dto.info.ImageInfo;
+import com.msa.community_service.domain.community.dto.info.MemberLoginActive;
 import com.msa.community_service.domain.community.dto.request.CommunityListRequest;
 import com.msa.community_service.domain.community.dto.request.CreateCommunityRequest;
 import com.msa.community_service.domain.community.dto.request.UpdateCommunityRequest;
@@ -16,7 +17,10 @@ import com.msa.community_service.domain.community.exception.CommunityException;
 import com.msa.community_service.domain.community.repository.CommentRepository;
 import com.msa.community_service.domain.community.repository.CommunityRepository;
 import com.msa.community_service.domain.community.repository.ImageRepository;
+import com.msa.community_service.global.common.dto.Message;
 import lombok.RequiredArgsConstructor;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -77,10 +81,15 @@ public class CommunityServiceImpl implements CommunityService {
     }
 
     @Override
-    public void updateCommunity(Long communityId, UpdateCommunityRequest request) {
+    public void updateCommunity(Long communityId, UpdateCommunityRequest request, MemberLoginActive memberLoginActive) {
         Community community = communityRepository.findById(communityId)
                 .orElseThrow(() -> new CommunityException(CommunityErrorCode.NOT_EXIST_COMMUNITY));
 
+        // 작성자 본인이 아닌 경우
+        if (memberLoginActive.role().name().equals("USER") && community.getWriterId() != memberLoginActive.id()){
+            System.out.println("권한이 없습니다.");
+            return;
+        }
         // 제목, 내용 수정
         community.update(request.title(), request.content());
 
@@ -98,24 +107,29 @@ public class CommunityServiceImpl implements CommunityService {
                         .community(community).url(imageInfo.url()).build());
                 continue;
             }
-
             // 기존에 있던 항목 map에서 제거 >> 이후 해당 map에 있는 값들을 한번에 삭제할 예정
             if (map.containsKey(imageId)) {
                 map.remove(imageId);
             }
         }
-
         images.removeAll(map.values());
     }
 
+    @Transactional
     @Override
-    public void deleteCommunity(Long communityId) {
+    public void deleteCommunity(Long communityId, MemberLoginActive memberLoginActive) {
+        Community community = communityRepository.findById(communityId)
+                .orElseThrow(() -> new CommunityException(CommunityErrorCode.NOT_EXIST_COMMUNITY));
+
+        // 작성자 본인이 아닌 경우
+        if (memberLoginActive.role().name().equals("USER") && community.getWriterId() != memberLoginActive.id()){
+            System.out.println("권한이 없습니다.");
+            return;
+        }
         // Image 삭제
         imageRepository.deleteByCommunityId(communityId);
-
         // Comment 삭제
         commentRepository.deleteByCommunityId(communityId);
-
         // Community 삭제
         communityRepository.deleteById(communityId);
     }
